requirements.txt
selenium
requests
pytest
appium-python-client
